V5.0
This commit includes Version 5 of the model, which is independent of
the h5py database creation tool. The images are preprocessed using a
new preprocess.py file into JPG’s without data augmentation. The
train.py file utilizes the image_preloader function to read all the
images, which are listed in a file in the scratch space using a newly
introduced dbfile. All the processors are optimized to make use of
multiple cores and gpu’s of the system they are being run on.
Additionally, this function allows the model to be scaled up to run on
more powerful hardware easily.